requirements.txt
elasticsearch[async]~=8.16.0
pytelegrambotapi~=4.25.0
loguru~=0.7.3
python-dotenv~=1.0.1
fastapi~=0.115.0
pydantic~=2.10.4
requests~=2.32.3
uvicorn~=0.27.1
gunicorn~=23.0.0
python-multipart~=0.0.20
IDU-config~=1.0.2
python-docx~=1.1.2
tqdm~=4.67.1
pandas~=2.2.3
websockets~=15.0.1